requirements.txt
fasttext==0.9.2
pyyaml==6.0.1
